@@ -10,17 +10,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.11",
+    "@angular/animations": "^14.1.1",
     "@angular/cdk": "^13.3.9",
-    "@angular/common": "~13.3.11",
-    "@angular/compiler": "~13.3.11",
-    "@angular/core": "~13.3.11",
+    "@angular/common": "^14.1.1",
+    "@angular/compiler": "^13.3.11",
+    "@angular/core": "^14.1.1",
     "@angular/flex-layout": "^12.0.0-beta.35",
-    "@angular/forms": "~13.3.11",
+    "@angular/forms": "^14.1.1",
     "@angular/material": "^13.3.9",
-    "@angular/platform-browser": "~13.3.11",
-    "@angular/platform-browser-dynamic": "~13.3.11",
-    "@angular/router": "~13.3.11",
+    "@angular/platform-browser": "^14.1.1",
+    "@angular/platform-browser-dynamic": "^14.1.1",
+    "@angular/router": "^14.1.1",
     "@zxing/browser": "0.0.9",
     "@zxing/library": "^0.18.6",
     "@zxing/ngx-scanner": "^3.2.0",
@@ -35,9 +35,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.9",
-    "@angular/cli": "~13.3.9",
-    "@angular/compiler-cli": "~13.3.11",
+    "@angular-devkit/build-angular": "^14.1.1",
+    "@angular/cli": "^14.1.1",
+    "@angular/compiler-cli": "^14.1.1",
     "@types/jasmine": "~3.8.0",
     "@types/node": "^12.11.1",
     "jasmine-core": "~3.8.0",
