@@ -1,5 +1,5 @@
 import {Component, OnDestroy, OnInit} from '@angular/core';
-import {FormBuilder, FormControl, FormGroup, Validators} from "@angular/forms";
+import {UntypedFormBuilder, UntypedFormControl, UntypedFormGroup, Validators} from "@angular/forms";
 import {MustMatch} from "../../helpers/must-match.validator";
 import {UiService} from "../../services/ui.service";
 import {AuthService} from "../../services/auth.service";
@@ -18,25 +18,25 @@ export class RegisterComponent implements OnInit, OnDestroy {
   passwordVisible = false;
   passwordConfirmVisible = false;
 
-  registerForm!: FormGroup;
+  registerForm!: UntypedFormGroup;
   isSubmitted = false;
 
   subscriptions: Subscription[] = []
 
   constructor(
-    private formBuilder: FormBuilder,
+    private formBuilder: UntypedFormBuilder,
     private ui: UiService,
     private auth: AuthService) { }
 
   ngOnInit(): void {
     this.registerForm = this.formBuilder.group({
-      'fullName': new FormControl(null, [Validators.required]),
-      'mobile': new FormControl(null,
+      'fullName': new UntypedFormControl(null, [Validators.required]),
+      'mobile': new UntypedFormControl(null,
         [Validators.required, Validators.pattern("^[0-9]*$"),
           Validators.minLength(10), Validators.maxLength(10)]),
-      'email': new FormControl(null, [Validators.required, Validators.email]),
-      'password': new FormControl(null, [Validators.required, Validators.minLength(8)]),
-      'passwordConfirm': new FormControl(null, [Validators.required])
+      'email': new UntypedFormControl(null, [Validators.required, Validators.email]),
+      'password': new UntypedFormControl(null, [Validators.required, Validators.minLength(8)]),
+      'passwordConfirm': new UntypedFormControl(null, [Validators.required])
     }, {
       validator: MustMatch('password', 'passwordConfirm')
     });
